rename directorySharedTypes.ts to globalTypes.ts
