+++ packages/vue-language-core/src/index.ts
@@ -8,7 +8,7 @@ export * from './utils/ts';
 export * from './utils/parseSfc';
 
 export * as scriptRanges from './parsers/scriptRanges';
-export * as sharedTypes from './utils/directorySharedTypes';
+export * as sharedTypes from './utils/globalTypes';
 
 export * from '@volar/language-core';
 export * from '@volar/source-map';
