+++ packages/vue-language-core/src/utils/globalTypes.ts
